requirements.txt
Django==5.1.4
django_imagekit==5.0.0
django_simple_captcha==0.6.2
